Added test step to circleci, moved shims to test folder, remove jest section from package.config
