--- src/react-raf-shim.js
@@ -1,5 +1,0 @@
-const raf = global.requestAnimationFrame = (cb) => {
-  setTimeout(cb, 0)
-}
-
-export default raf
