+++ src/views/home/home.spec.js
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import raf from '../../react-raf-shim';
+import raf from '../../../test/shims/react-raf-shim';
 import { shallow, configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
